refactor: move initialize PX to shared/PX
